graphics and ux improvements

+++ frontend/src/app/clientcard/clientcard.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, Input, OnInit, Output, EventEmitter } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { Client } from 'src/model/client.modul';
 import { ClientsService } from '../clients.service';
@@ -11,12 +11,19 @@ import { ClientsService } from '../clients.service';
 export class ClientcardComponent implements OnInit {
   @Input()
   client!: Client;
+  @Output() update: EventEmitter<string> = new EventEmitter();
   constructor(private clientService: ClientsService) {}
+
+  isModalVisible: boolean = false;
 
   //delete clients
   deleteClient(id: number) {
-    this.clientService.deleteClient(id);
-    window.location.reload();
+    if (
+      confirm(`Vuoi davvero eliminare il cliente ${this.client.business_name}`)
+    ) {
+      this.clientService.deleteClient(id);
+      this.update.emit('update');
+    }
   }
 
   ngOnInit(): void {}
